moved validation into seperate file
